Move code for running the tests to RunTests.js
